requirements.txt
# requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.35
python-dotenv==1.0.1
faiss-cpu==1.7.4
pypdf==3.8.1
transformers==4.35.0
sentence-transformers==2.2.2
torch==2.0.1
scikit-learn==1.2.2
numpy==1.23.5
openlit
chromadb